fatcorkbackend/requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==41.0.2
Django==4.2.3
django-encrypted-model-fields==0.6.5
django-phonenumber-field==7.1.0
django-taggit==4.0.0
phonenumberslite==8.13.16
psycopg==3.1.9
psycopg-binary==3.1.9
pycparser==2.21
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
